requirements.txt
gradio>=3.4.0
farm-haystack>=1.8.0